+++ src/__mocks__/vscode.ts
@@ -37,6 +37,11 @@ export class CodeActionKind {
   constructor(readonly value: string) {}
 }
 
+export enum CodeActionTriggerKind {
+  Invoke = 1,
+  Automatic = 2,
+}
+
 export class Diagnostic {
   code?: string
   source?: string
@@ -142,9 +147,13 @@ export const workspace = {
     ),
   })),
   getWorkspaceFolder: jest.fn(),
+  get isTrusted() {
+    return true
+  },
   onDidChangeConfiguration: jest.fn(),
   onDidChangeTextDocument: jest.fn(),
   onDidCloseTextDocument: jest.fn(),
+  onDidGrantWorkspaceTrust: jest.fn(),
   textDocuments: [] as any,
 
   _setConfiguration: (conf: MockedConf = {}) => {